Разбивает код на модули и добавляет сборку с помощью webpack
