+++ public/index.html
@@ -41,7 +41,7 @@
             </section>
         </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
 </body>
 
 </html>
